Refactor PublicProfile to take profileData as props
